+++ python_ta/contracts/__init__.py
@@ -266,11 +266,8 @@ def check_type_strict(argname: str, value: Any, expected_type: type) -> None:
         - float vs. int
         - bool vs. int
     """
-    if (
-        type(value) is int
-        and expected_type is float
-        or type(value) is bool
-        and expected_type is int
+    if (type(value) is int and expected_type is float) or (
+        type(value) is bool and expected_type is int
     ):
         raise TypeError(f"type of {argname} must be {expected_type}; got {value} instead")
     check_type(argname, value, expected_type)